ALMOOSA-1007 ALMOOSA-1008 ALMOOSA-1009 ALMOOSA-1010 ALMOOSA-1011 Services & Packages Module

+++ src/pages/payment-success.vue
@@ -15,9 +15,11 @@ export default {
     ...mapActions("appointment", ["setSelectedAppointment"]),
     showModal() {
       let booking = userService.getBooking();
-      booking.status = "confirmed";
-      this.setSelectedAppointment(booking);
-      userService.removeBooking();
+      if (booking) {
+        booking.status = "confirmed";
+        this.setSelectedAppointment(booking);
+        userService.removeBooking();
+      }
       this.setLoadingState(false);
       this.successIconModal(
         this.$t("selectPaymentMethod.paymentSuccessful"),
